fix(ckeditor): fixes basepath issues on some systems
As the CKeditor library is not a real AMD library, its execution may not be
controlled as well by RequireJS, causing a race condition with the module
that sets the global `CKEDITOR_BASEPATH` constant.

This renames the module `elgg/ckeditor` to `elgg-ckeditor` which eliminates
the need for setting `CKEDITOR_BASEPATH` at all. With the new filename,
the CKeditor library is no longer confused when sniffing for its resources
folder.

For BC, the now-unneeded modules have been left in place and deprecated, and
the global `CKEDITOR_BASEPATH` is set via inline code in `elgg.js`.

Fixes #10724
