+++ mod/ckeditor/start.php
@@ -14,18 +14,17 @@ function ckeditor_init() {
 	elgg_extend_view('elgg/wysiwyg.css', 'elements/reset.css', 100);
 	elgg_extend_view('elgg/wysiwyg.css', 'elements/typography.css', 100);
 
-	elgg_define_js('ckeditor', array(
-		'deps' => ['elgg/ckeditor/set-basepath'],
+	elgg_define_js('ckeditor/ckeditor', array(
 		'exports' => 'CKEDITOR',
 	));
 	elgg_define_js('jquery.ckeditor', array(
-		'deps' => array('jquery', 'ckeditor'),
+		'deps' => array('jquery', 'ckeditor/ckeditor'),
 		'exports' => 'jQuery.fn.ckeditor',
 	));
 
-	// need to set basepath early
-	elgg_extend_view('elgg.js', 'elgg/ckeditor/set-basepath.js');
-	
+	// TODO remove in 3.0
+	elgg_extend_view('elgg.js', 'elgg/ckeditor/set-basepath', 499);
+
 	elgg_extend_view('input/longtext', 'ckeditor/init');
 
 	elgg_register_plugin_hook_handler('register', 'menu:longtext', 'ckeditor_longtext_menu');
